Conflicts:
	cocos/editor-support/cocostudio/CCActionObject.cpp
	samples/Cpp/TestCpp/Classes/PhysicsTest/PhysicsTest.h

+++ cocos/scripting/lua/bindings/LuaBasicConversions.cpp
@@ -1203,7 +1203,7 @@ void array_to_luaval(lua_State* L,Array* inValue)
         if (nullptr == obj)
             continue;
         
-        uint32_t typeId = typeid(*obj).hash_code();
+        long typeId = typeid(*obj).hash_code();
         auto iter = g_luaType.find(typeId);
         if (g_luaType.end() != iter)
         {
@@ -1292,7 +1292,7 @@ void dictionary_to_luaval(lua_State* L, Dictionary* dict)
         if (NULL == element)
             continue;
         
-        uint32_t typeId = typeid(element->getObject()).hash_code();
+        long typeId = typeid(element->getObject()).hash_code();
         
         auto iter = g_luaType.find(typeId);
         if (g_luaType.end() != iter)
